requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
aspose-cells-python==22.7.0
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==4.2.2
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc8
flake8==4.0.1
idna==3.3
lxml==4.9.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-cache==0.9.5
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
url-normalize==1.4.3
urllib3==1.26.10
xlrd==2.0.1
